lab2+lab3
lab3: lower the toxicity of our instruction fine-tuned model from Lab 2, that was the output of Lab 2. That'll be the input to Lab 3 here. And we will lower the toxicity using a hate speech reward model where we want to optimize for not hate. We'll be using PPO that you learned about in the lesson, and we will wrap everything up with quantitative and then qualitative comparison of our detoxification process.